feat: set integer size for FTorch ftorch_int
Integer size for FTorch can now be controlled using parameter
`ftorch_int`.

Currently, the default is set to `int32` but this could be changed if
required.

This fixes an issue on windows with the latest version of ifx/ifort
which builds with a different integer type leading to the following
error:

```
error #6284: There is no matching specific function for this generic function reference.   [TORCH_TENSOR_FROM_ARRAY]
   in_tensors(1) = torch_tensor_from_array(in_data, tensor_layout, torch_kCPU)
-------------------^
```