Bugfix: explicitly move loaded net to required device in multigpu python in example 3
If the net is saved from a particular cuda device it will be re-loaded
back to that cuda device. This causes issues if we are placing the
tensors on a different device, as is the case in the multigpu inference.
To fix this explicitly move the loaded net to the device.

We should look a little deeper about how this works on the Fortran/C++
side and verify where things are saved from/loaded to.