Fix bug in implementation of wrapper function `torch_from_blob` (#365)
* Roll back to previous approach (6deab9a)

* In the current implementation (864b0eb) the device is passed
explicitly to TensorOptions object and then the call to torch::from_blob
is made passing that object. This means that in the case of a GPU device
the data is supposed to be residing on the GPU therefore throwing
[ERROR]: The specified pointer resides on host memory and is not
registered with any CUDA device.

* Now device is not passed explicitly via the options. The tensor is
then moved to the target GPU device