Fix RPATH for installed libftorch to ensure it can be linked by downstream targets (fixes #435)
-------
Summary
-------
* updated toplevel cmakelists to install rpath with Torch dependency
* updated changelog
* updated hpc docs

-------
Details
-------

During the cmake **installation** phase for FTorch, the produced shared
library `libftorch.so` had a `RUNPATH` (inspected with `ldd`) that did
not contain a path to the Torch library. This meant that downstream
targets would need to modify `LD_LIBARY_PATH` to contain the path to the
Torch library in order to link with FTorch (which is dependent on
Torch). By modifying the `RPATH` used in the cmake installation phase
to contain the Torch library path, `libftorch.so` produced during this
phase contains the correct `RUNPATH` pointing to the directory where
the Torch library is.

This problem did not show up during testing because the binaries
compiled for e.g., integration tests were linked against the `libftorch.so`
that is produced during the cmake **build** phase. The `libftorch.so`
produced during the cmake build phase has the **correct** `RUNPATH` for
the Torch library, and therefore FTorch could find the Torch shared
libraries (e.g., `libc10.so`) on which FTorch depends.